ChatBook/requirements.txt
python-dotenv
gradio
langchain
langchain-community
langchain-openai
langdetect==1.0.9